requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asyncer==0.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
Django==4.0.5
fastapi==0.72.0
filelock==3.7.1
filetype==1.0.9
Flask==2.1.2
flatbuffers==2.0
frozenlist==1.3.0
gdown==4.4.0
h11==0.13.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.38.1
MarkupSafe==2.1.1
multidict==6.0.2
networkx==2.8.3
numba==0.55.2
numpy==1.22.3
onnxruntime==1.10.0
packaging==21.3
Pillow==9.0.1
protobuf==4.21.1
pydantic==1.9.1
PyMatting==1.1.5
pyparsing==3.0.9
PySocks==1.7.1
python-multipart==0.0.5
PyWavelets==1.3.0
rembg==2.0.21
requests==2.28.0
requests-toolbelt==0.9.1
scikit-image==0.19.1
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starlette==0.17.1
tifffile==2022.5.4
tqdm==4.62.3
transfer.sh==2.0.1
typing-extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
uvicorn==0.17.0
watchdog==2.1.7
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
